container/requirements.txt
Flask
pybase64
jsonpickle
Pillow
numpy
tqdm
opencv-python
fastapi
tensorflow
tensorflow-hub